3-project-flink-plus-tableflow/requirements.txt
confluent-kafka>=2.8.0
python-dotenv
Faker>=19.6.2
pytest
pytest-timeout
pytest-asyncio
jsonschema>=4.0.0
requests>=2.25.0
httpx
cachetools
authlib
referencing
fastavro
duckdb
numpy
pandas